+++ src/aui-datepicker/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-2103](https://issues.liferay.com/browse/AUI-2103) Date Picker fire a selectedDateChange event when the date value changes and Time picker fires selectedTimeChange when the time value changes
 * [AUI-2086](https://issues.liferay.com/browse/AUI-2086) DatePickerNative throws error when destroyed
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)